renamed game_frontend -> old_frontend
