+++ src/game.rs
@@ -13,10 +13,10 @@ use crate::entity::object::Object;
 use crate::entity::player::PlayerCtrl;
 use crate::ui::color_palette::ColorPalette;
 use crate::ui::frontend::render_world;
-use crate::ui::game_frontend::{handle_meta_actions, process_visual_feedback};
 use crate::ui::game_input::{GameInput, PlayerInput};
 use crate::ui::gui::render_gui;
 use crate::ui::menu::{display_main_menu, MenuInstance};
+use crate::ui::old_frontend::{handle_meta_actions, process_visual_feedback};
 use rltk::{GameState as Rltk_GameState, Rltk, RGB};
 use std::error::Error;
 use std::fs::{self, File};
