+++ src/ui/dialog.rs
@@ -6,10 +6,8 @@ use tcod::{chars, console::*, input::Key};
 
 use crate::core::game_objects::GameObjects;
 use crate::core::game_state::GameState;
-use crate::ui::game_frontend::{
-    GameFrontend, BAR_WIDTH, PANEL_HEIGHT, SCREEN_HEIGHT, SCREEN_WIDTH,
-};
 use crate::ui::game_input::GameInput;
+use crate::ui::old_frontend::{GameFrontend, BAR_WIDTH, PANEL_HEIGHT, SCREEN_HEIGHT, SCREEN_WIDTH};
 use crate::util::modulus;
 
 // message box measurements
